the method widens doe alloc and copy sizes to size_t before the int overflow
CodeQL (default setup, threat model remote) flagged 16
cpp/integer-multiplication-cast-to-long alerts in doe.c: int*int products used as
allocation or copy sizes that overflow in int before the implicit widen to size_t. A
crafted or oversized host profile could wrap a product to a small value and
under-allocate a LoRA / KV / RoPE buffer, then write past its end.

Each flagged size expression casts its leading operand to size_t so the product computes
wide: LoRA experts dim*rank / rank*dim in calloc/fwrite/fread (including the readback
compare, which was an int product cast after the fact); alloc_infer host_heads*head_dim,
host_heads*max_seq, host_n_layers*max_seq*kd, max_seq*half; the KV-reset memset in chat,
now matching the already-cast kv_bytes and the sibling reset path.

Numerically inert for in-range sizes. Proof (neo): make test -> 113/113; cc -O2 -Wall
-Wextra -c doe.c clean apart from the one pre-existing unused-'hs' warning. Canon fix —
the yent-inference vendored copy was already patched by Codex; this closes the source.
DOELOG started here (the log was missing).

Quote: "In the multitude of counsellors there is safety." — Proverbs 11:14
Method: the Method counts its parliament's buffers in the wide word, so no product wraps before the vote.

Co-authored by Claude Opus 4.8 (Arianna Method, neo) <theariannamethod@gmail.com>, Coordinated with maintainer @iamolegataeff

+++ doe.c
@@ -1746,8 +1746,8 @@ static void env_scan(Environment *env, const char *self_src) {
  * the weights are substrate. DOE is the architecture.
  * ═══════════════════════════════════════════════════════════════════════════════ */
 static void init_lora_expert(LoraExpert *e, int dim, int rank, float freq) {
-    e->lora_A = calloc(dim * rank, sizeof(float));
-    e->lora_B = calloc(rank * dim, sizeof(float));
+    e->lora_A = calloc((size_t)dim * rank, sizeof(float));
+    e->lora_B = calloc((size_t)rank * dim, sizeof(float));
     float scale = 0.02f / sqrtf((float)rank);
     for (int i = 0; i < dim*rank; i++) e->lora_A[i] = rand_normal() * scale;
     for (int i = 0; i < rank*dim; i++) e->lora_B[i] = rand_normal() * scale;
@@ -2659,8 +2659,8 @@ static void mycelium_save(GGUFIndex *ps, int step, float fitness) {
             if (ex->alive) {
                 fwrite(&ex->vitality, 4, 1, f);
                 fwrite(&ex->frequency, 4, 1, f);
-                fwrite(ex->lora_A, sizeof(float), dim * rank, f);
-                fwrite(ex->lora_B, sizeof(float), rank * dim, f);
+                fwrite(ex->lora_A, sizeof(float), (size_t)dim * rank, f);
+                fwrite(ex->lora_B, sizeof(float), (size_t)rank * dim, f);
             }
         }
     }
@@ -2719,14 +2719,14 @@ static int mycelium_load(GGUFIndex *ps, uint64_t target_fp) {
             int alive; if (fread(&alive, 4, 1, f) != 1) { fclose(f); return 0; }   /* C-2 */
             if (alive) {
                 if (!ex->alive) {
-                    ex->lora_A = calloc(dim * rank, sizeof(float));
-                    ex->lora_B = calloc(rank * dim, sizeof(float));
+                    ex->lora_A = calloc((size_t)dim * rank, sizeof(float));
+                    ex->lora_B = calloc((size_t)rank * dim, sizeof(float));
                     if (!ex->lora_A || !ex->lora_B) { fclose(f); return 0; }   /* C-2/F-1: OOM */
                 }
                 ex->alive = 1;
                 if (fread(&ex->vitality, 4, 1, f) != 1 || fread(&ex->frequency, 4, 1, f) != 1) { fclose(f); return 0; }   /* C-2 */
-                if (fread(ex->lora_A, sizeof(float), dim * rank, f) != (size_t)(dim * rank) ||
-                    fread(ex->lora_B, sizeof(float), rank * dim, f) != (size_t)(rank * dim)) {
+                if (fread(ex->lora_A, sizeof(float), (size_t)dim * rank, f) != (size_t)dim * rank ||
+                    fread(ex->lora_B, sizeof(float), (size_t)rank * dim, f) != (size_t)rank * dim) {
                     fclose(f); return 0; /* D-L3: truncated spore — bail, don't load garbage experts */
                 }
             } else if (ex->alive) {
@@ -2773,9 +2773,9 @@ static InferState alloc_infer(GGUFIndex *ps, int max_seq) {
     int H = ps->host_hidden;
     s.max_seq = max_seq;
     s.x = calloc(D, 4); s.xb = calloc(D, 4); s.xb2 = calloc(D, 4);
-    s.q = calloc(ps->host_heads * ps->host_head_dim, 4);
+    s.q = calloc((size_t)ps->host_heads * ps->host_head_dim, 4);
     s.k = calloc(kd, 4); s.v = calloc(kd, 4);
-    s.att = calloc(ps->host_heads * max_seq, 4);
+    s.att = calloc((size_t)ps->host_heads * max_seq, 4);
     s.logits = calloc(ps->host_vocab, 4);
     s.hb = calloc(H, 4); s.hb2 = calloc(H * 2, 4); /* *2 for fused gate_up */
     s.expert_out = calloc(D, 4);
@@ -2787,8 +2787,8 @@ static InferState alloc_infer(GGUFIndex *ps, int max_seq) {
     if (posix_memalign((void**)&s.key_cache,   (size_t)getpagesize(), kv_bytes) ||
         posix_memalign((void**)&s.value_cache, (size_t)getpagesize(), kv_bytes)) {
         fprintf(stderr, "[doe] KV posix_memalign failed — falling back to calloc\n");
-        s.key_cache = calloc(ps->host_n_layers * max_seq * kd, 4);
-        s.value_cache = calloc(ps->host_n_layers * max_seq * kd, 4);
+        s.key_cache = calloc((size_t)ps->host_n_layers * max_seq * kd, 4);
+        s.value_cache = calloc((size_t)ps->host_n_layers * max_seq * kd, 4);
     } else {
         memset(s.key_cache, 0, kv_bytes);
         memset(s.value_cache, 0, kv_bytes);
@@ -2800,8 +2800,8 @@ static InferState alloc_infer(GGUFIndex *ps, int max_seq) {
 #endif
     }
     int half = ps->host_head_dim / 2;
-    s.cos_cache = calloc(max_seq * half, 4);
-    s.sin_cache = calloc(max_seq * half, 4);
+    s.cos_cache = calloc((size_t)max_seq * half, 4);
+    s.sin_cache = calloc((size_t)max_seq * half, 4);
     float rope_theta = ps->rope_theta;
     for (int p = 0; p < max_seq; p++)
         for (int i = 0; i < half; i++) {
@@ -3636,8 +3636,8 @@ static void chat(GGUFIndex *ps) {
 
         /* Reset KV cache */
         int kd = ps->host_kv_heads * ps->host_head_dim;
-        memset(is.key_cache, 0, ps->host_n_layers * max_seq * kd * 4);
-        memset(is.value_cache, 0, ps->host_n_layers * max_seq * kd * 4);
+        memset(is.key_cache, 0, (size_t)ps->host_n_layers * max_seq * kd * 4);
+        memset(is.value_cache, 0, (size_t)ps->host_n_layers * max_seq * kd * 4);
 
         /* Wrap input in chat template (auto-detected from GGUF chat_template) */
         char wrapped[8192];  /* F-12: match input[8192] so the chat template's closing tags are never silently truncated */
